Update Chart.php
I have this change working for me, so I figure I would share it. Basically, it will allow users to set the Min/Max value of the yAxis in their Chart. This change goes together with more changes in the Chart.php file located in the Writer/Excel2007 folder.

Usage can be done as in the following:
$yAxisMinValue = 90;
$yAxisMaxValue = 500;

//	Create the chart
$chart = new PHPExcel_Chart(
	'chart1',		// name
	$title,			// title
	$legend,		// legend
	$plotarea,		// plotArea
	true,			// plotVisibleOnly
	0,				// displayBlanksAs
	NULL,			// xAxisLabel
	$yAxisLabel,		//
        $yAxisMinValue,         // yAxisMinvalue
        $yAxisMaxValue,         // yAxisMaxValue
);